chore: update imports, rename declarationData to declaration
