+++ packages/client/src/v2-events/features/events/AdvancedSearch/TabSearch.tsx
@@ -17,7 +17,7 @@ import { Accordion } from '@opencrvs/components'
 import { Icon } from '@opencrvs/components/lib/Icon'
 import { Button } from '@opencrvs/components/lib/Button'
 import { EventConfig, FieldValue } from '@opencrvs/commons/client'
-import { FormFieldGenerator } from '@client/v2-events/components/forms/FormFieldGenerator'
+import { FormFieldGenerator } from '@client/v2-events/components/forms/FormFieldGenerator/FormFieldGenerator'
 import { getAllUniqueFields } from '@client/v2-events/utils'
 import { ROUTES } from '@client/v2-events/routes'
 import { flattenFieldErrors, getAdvancedSearchFieldErrors } from './utils'
