+++ packages/client/src/v2-events/features/events/components/Pages.tsx
@@ -19,7 +19,7 @@ import {
   PageConfig
 } from '@opencrvs/commons/client'
 import { MAIN_CONTENT_ANCHOR_ID } from '@opencrvs/components/lib/Frame/components/SkipToContent'
-import { FormFieldGenerator } from '@client/v2-events/components/forms/FormFieldGenerator'
+import { FormFieldGenerator } from '@client/v2-events/components/forms/FormFieldGenerator/FormFieldGenerator'
 import { usePagination } from '@client/v2-events/hooks/usePagination'
 import { VerificationWizard } from './VerificationWizard'
 import { FormWizard } from './FormWizard'
@@ -87,8 +87,8 @@ export function Pages({
 
   const fields = (
     <FormFieldGenerator
+      declaration={declaration}
       eventConfig={eventConfig}
-      eventDeclarationData={declaration}
       fields={page.fields}
       formData={form}
       id="locationForm"
