+++ packages/client/src/v2-events/features/events/registered-fields/Data.stories.tsx
@@ -14,7 +14,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { noop } from 'lodash'
 import { FieldType, tennisClubMembershipEvent } from '@opencrvs/commons/client'
-import { FormFieldGenerator } from '@client/v2-events/components/forms/FormFieldGenerator'
+import { FormFieldGenerator } from '@client/v2-events/components/forms/FormFieldGenerator/FormFieldGenerator'
 import { TRPCProvider } from '@client/v2-events/trpc'
 
 const meta: Meta<typeof FormFieldGenerator> = {
@@ -42,13 +42,13 @@ export const DataDisplay: StoryObj<typeof FormFieldGenerator> = {
   render: function Component() {
     return (
       <StyledFormFieldGenerator
-        eventConfig={tennisClubMembershipEvent}
-        eventDeclarationData={{
+        declaration={{
           'applicant.firstname': 'Tanya',
           'applicant.id': '2370934578',
           'applicant.surname': 'McQuaid',
           'applicant.dob': '1975-01-02'
         }}
+        eventConfig={tennisClubMembershipEvent}
         fields={[
           {
             id: 'storybook.data',
@@ -107,12 +107,12 @@ export const DataDisplayWithConditionallyHiddenFields: StoryObj<
   render: function Component() {
     return (
       <StyledFormFieldGenerator
-        eventConfig={tennisClubMembershipEvent}
-        eventDeclarationData={{
+        declaration={{
           'recommender.none': true,
           'recommender.firstname': 'John',
           'applicant.firstname': 'Rasheed'
         }}
+        eventConfig={tennisClubMembershipEvent}
         fields={[
           {
             id: 'storybook.data',
@@ -161,12 +161,12 @@ export const DataDisplayWithConditionallyShownFields: StoryObj<
   render: function Component() {
     return (
       <StyledFormFieldGenerator
-        eventConfig={tennisClubMembershipEvent}
-        eventDeclarationData={{
+        declaration={{
           'recommender.none': false,
           'recommender.firstname': 'John',
           'applicant.firstname': 'Rasheed'
         }}
+        eventConfig={tennisClubMembershipEvent}
         fields={[
           {
             id: 'storybook.data',
