+++ packages/events/src/router/user/index.ts
@@ -10,9 +10,9 @@
  */
 
 import { z } from 'zod'
+import { TRPCError } from '@trpc/server'
 import { router, publicProcedure } from '@events/router/trpc'
 import { getUsersById } from '@events/service/users/users'
-import { TRPCError } from '@trpc/server'
 
 export const userRouter = router({
   get: publicProcedure.input(z.string()).query(async ({ input }) => {
